+++ packages/dev/core/src/Maths/index.ts
@@ -3,3 +3,4 @@ export * from "./math.functions";
 export * from "./math.polar";
 export * from "./math";
 export * from "./sphericalPolynomial";
+export * from "./halton2DSequence";
