+++ packages/dev/core/src/Misc/index.ts
@@ -66,4 +66,3 @@ export * from "./copyTextureToTexture";
 export * from "./dumpTools";
 export * from "./greasedLineTools";
 export * from "./equirectangularCapture";
-export * from "./halton2DSequence";
